+++ src/components/index.js
@@ -1,0 +1,10 @@
+export { default as Card } from './Card';
+export { default as CartCard } from './CartCard';
+export { default as CheckBox } from './CheckBox';
+export { default as RadioInput } from './RadioInput';
+export { default as ClearButton } from './ClearButton';
+export { default as InputRange } from './InputRange';
+export { default as Nav } from './Nav';
+export { default as Filter } from './Filter';
+export { default as Footer } from './Footer';
+export { default as PriceDetails } from './PriceDetails';
