+++ src/pages/auth/Login.jsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
-import './Auth.css';
 import { useDispatch } from 'react-redux';
 import { toast } from 'react-toastify';
+import './Auth.css';
 import { loginUser } from '../../redux/features/authSlice';
 import { getWishlistThunk } from '../../redux/features/wishListSlice';
 import { getCart } from '../../redux/features/cartSlice';
