+++ src/pages/index.js
@@ -1,0 +1,6 @@
+export { default as Home } from './homepage/Home';
+export { default as Login } from './auth/Login';
+export { default as SignUp } from './auth/SignUp';
+export { default as Cart } from './cart/Cart';
+export { default as Products } from './products/Products';
+export { default as WishList } from './wishList/WishList';
